+++ lib/Constants.js
@@ -9,7 +9,7 @@ const CLIENT_STATUS = new Enum(['NONE', 'START', 'STARTING', 'STARTED',
 
 const MESSAGE_TYPE = new Enum(['START', 'STARTING', 'STARTED',
     'CONNECTING', 'CONNECTED', 'PAUSE', 'PAUSING', 'PAUSED',
-    'RESUME', 'RESUMING', 'RESUMED', 'CLOSE', 'CLOSING', 'CLOSED', 'RECEIPT', 'COMMITTED'], {
+    'RESUME', 'RESUMING', 'RESUMED', 'CLOSE', 'CLOSING', 'CLOSED', 'RECEIPT', 'COMMITTED' ,'ERROR'], {
     freez: true,
     name: 'MESSAGE_TYPE'
 });
