+++ lib/LambdaStreamClient.js
@@ -9,12 +9,12 @@ const MESSAGE_TYPE = require('./Constants').MESSAGE_TYPE;
 
 
 class LambdaStreamClient extends EventEmitter {
-    constructor(endPoint, lChainId, authorization) {
+    constructor(endPoint, lChainId, accessKey) {
         super();
 
         this.endPoint = endPoint;
         this.lChainId = lChainId;
-        this.authorization = authorization;
+        this.accessKey = accessKey;
 
         this.wsClient = null;
 
@@ -46,7 +46,7 @@ class LambdaStreamClient extends EventEmitter {
                 }
 
                 self.processMutex = false;
-            }, 1000);
+            }, 3000);
         }
     }
 
@@ -131,11 +131,10 @@ class LambdaStreamClient extends EventEmitter {
     }
 
     async connect() {
-        //TODO:.. add authorization
         let options = {
             headers: {
                 lChainId: this.lChainId,
-                authorization: this.authorization
+                accessKey: this.accessKey
             }
         };
         /////////////////////////////
